clean process exit without waiting on condition variables (..)
This possibly fixes issue #2, see
https://github.com/matgoebl/kogmo-rtdb/issues/2 .

The goal of this commit is to programmatically avoid the case that a process
is shut down while one of its threads is waiting on a condition variable.
According to the POSIX standard, this is undefined behaviour and the condition
variable may be in an inconsistent state afterwards.

The following changes were made:
- signal handling (thus also exit handling) performed in an extra thread
- 'terminate' flag in process-local data to notify threads that they should
  terminate, i.e. stop waiting on condition variables etc.
- periodically timeout while waiting on a condition variable, so the
  termination flag can be checked (timeouts every 100ms)
- optional 500ms wait in kogmo_rtdb_ipc_disconnect to allow all threads to stop
  waiting on condition variables (set via KOGMO_RTDB_DISCONNECT_FLAG_WAIT)
- make the rtdb mutexes robust, just to be sure

As a side-effect, condition variable handling is now also robust against
spurious wakeups by checking the time the last condition variable signalling
was performed.